# docs/api/openapi.yml
components:
  schemas:
...
    AssignmentOut:
      properties:
        berth_id:
          examples:
          - berth-001
          title: Berth Id
          type: string
        user_id:
          examples:
          - user-001
          title: User Id
          type: string
      required:
      - berth_id
      - user_id
      title: AssignmentOut
      type: object
...
    BerthOut:
      properties:
        assignment:
          anyOf:
          - $ref: '#/components/schemas/AssignmentOut'
          - type: 'null'
        battery_pct:
          anyOf:
          - examples: &id005
            - 87
            type: integer
          - type: 'null'
          examples: *id005
          title: Battery Pct
        berth_id:
          examples:
          - berth-001
          title: Berth Id
          type: string
        depth_m:
          anyOf:
          - examples: &id006
            - 2.0
            type: number
          - type: 'null'
          examples: *id006
          title: Depth M
        dock_id:
          examples:
          - dock-a
          title: Dock Id
          type: string
        is_reserved:
          default: false
          title: Is Reserved
          type: boolean
        label:
          anyOf:
          - examples: &id007
            - A1
            type: string
          - type: 'null'
          examples: *id007
          title: Label
        last_updated:
          anyOf:
          - examples: &id008
            - '2026-05-03T14:30:00Z'
            format: date-time
            type: string
          - type: 'null'
          examples: *id008
          title: Last Updated
        length_m:
          anyOf:
          - examples: &id009
            - 8.5
            type: number
          - type: 'null'
          examples: *id009
          title: Length M
        sensor_raw:
          anyOf:
          - examples: &id010
            - 1234
            type: integer
          - type: 'null'
          examples: *id010
          title: Sensor Raw
        status:
          enum:
          - free
          - occupied
          title: Status
          type: string
        width_m:
          anyOf:
          - examples: &id011
            - 3.2
            type: number
          - type: 'null'
          examples: *id011
          title: Width M
      required:
      - berth_id
      - dock_id
      - status
      title: BerthOut
      type: object
...
    DockAdminOut:
      properties:
        dock_id:
          title: Dock Id
          type: string
        harbor_id:
          title: Harbor Id
          type: string
        name:
          title: Name
          type: string
      required:
      - dock_id
      - harbor_id
      - name
      title: DockAdminOut
      type: object
    DockCreate:
      properties:
        dock_id:
          maxLength: 64
          minLength: 1
          title: Dock Id
          type: string
        harbor_id:
          maxLength: 64
          minLength: 1
          title: Harbor Id
          type: string
        name:
          maxLength: 128
          minLength: 1
          title: Name
          type: string
      required:
      - dock_id
      - harbor_id
      - name
      title: DockCreate
      type: object
    DockOut:
      properties:
        dock_id:
          examples:
          - dock-a
          title: Dock Id
          type: string
        harbor_id:
          examples:
          - harbor-saltsjobaden
          title: Harbor Id
          type: string
        name:
          examples:
          - A Pier
          title: Name
          type: string
      required:
      - dock_id
      - harbor_id
      - name
      title: DockOut
      type: object
    DockPatch:
      properties:
        harbor_id:
          anyOf:
          - maxLength: 64
            minLength: 1
            type: string
          - type: 'null'
          title: Harbor Id
        name:
          anyOf:
          - maxLength: 128
            minLength: 1
            type: string
          - type: 'null'
          title: Name
      title: DockPatch
      type: object
    DockWithBerthsOut:
      properties:
        berths:
          default: []
          items:
            $ref: '#/components/schemas/BerthOut'
          title: Berths
          type: array
        dock_id:
          examples:
          - dock-a
          title: Dock Id
          type: string
        harbor_id:
          examples:
          - harbor-saltsjobaden
          title: Harbor Id
          type: string
        name:
          examples:
          - A Pier
          title: Name
          type: string
      required:
      - dock_id
      - harbor_id
      - name
      title: DockWithBerthsOut
      type: object
...
    GatewayCreate:
      properties:
        dock_id:
          maxLength: 64
          minLength: 1
          title: Dock Id
          type: string
        gateway_id:
          maxLength: 64
          minLength: 1
          title: Gateway Id
          type: string
        name:
          maxLength: 128
          minLength: 1
          title: Name
          type: string
      required:
      - gateway_id
      - dock_id
      - name
      title: GatewayCreate
      type: object
    GatewayCreatedOut:
      properties:
        gateway_id:
          title: Gateway Id
          type: string
        status:
          examples:
          - offline
          title: Status
          type: string
      required:
      - gateway_id
      - status
      title: GatewayCreatedOut
      type: object
    GatewayOut:
      properties:
        dock_id:
          examples:
          - dock-a
          title: Dock Id
          type: string
        gateway_id:
          examples:
          - gw-dock-a
          title: Gateway Id
          type: string
        last_seen:
          anyOf:
          - examples: &id012
            - '2026-05-03T14:30:00Z'
            format: date-time
            type: string
          - type: 'null'
          examples: *id012
          title: Last Seen
        name:
          examples:
          - Pier A gateway
          title: Name
          type: string
        provision_ttl_s:
          anyOf:
          - examples: &id013
            - 180
            type: integer
          - type: 'null'
          description: Per-gateway override; null falls back to ADOPTION_TTL
          examples: *id013
          title: Provision Ttl S
        status:
          enum:
          - online
          - offline
          title: Status
          type: string
      required:
      - gateway_id
      - dock_id
      - name
      - status
      title: GatewayOut
      type: object
    GatewayPatch:
      properties:
        dock_id:
          anyOf:
          - maxLength: 64
            minLength: 1
            type: string
          - type: 'null'
          description: Reassign to dock
          title: Dock Id
        name:
          anyOf:
          - maxLength: 128
            minLength: 1
            type: string
          - type: 'null'
          title: Name
        provision_ttl_s:
          anyOf:
          - maximum: 3600.0
            minimum: 10.0
            type: integer
          - type: 'null'
          title: Provision Ttl S
      title: GatewayPatch
      type: object
    GatewayPatchOut:
      properties:
        dock_id:
          title: Dock Id
          type: string
        gateway_id:
          title: Gateway Id
          type: string
        name:
          title: Name
          type: string
        provision_ttl_s:
          anyOf:
          - type: integer
          - type: 'null'
          title: Provision Ttl S
      required:
      - gateway_id
      - name
      - dock_id
      title: GatewayPatchOut
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
  securitySchemes:
    APIKeyCookie:
      in: cookie
      name: dockpulse_access
      type: apiKey
info:
  description: 'REST API for the DockPulse harbor berth availability system.


    The OpenAPI document at `/openapi.json` (and the YAML mirror in

    `docs/api/openapi.yml`) is generated directly from these route

    definitions and Pydantic models. The committed YAML is regenerated by

    `scripts/dump_openapi.py` and CI fails on drift.

    '
  title: DockPulse API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/admin/docks:
    get:
      operationId: adminListDocks
      parameters:
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/DockAdminOut'
                title: Response Adminlistdocks
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Docks
      tags:
      - admin
    post:
      operationId: adminCreateDock
      parameters:
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DockCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DockAdminOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Dock
      tags:
      - admin
  /api/admin/docks/{dock_id}:
    delete:
      operationId: adminDeleteDock
      parameters:
      - in: path
        name: dock_id
        required: true
        schema:
          title: Dock Id
          type: string
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Dock
      tags:
      - admin
    patch:
      operationId: adminPatchDock
      parameters:
      - in: path
        name: dock_id
        required: true
        schema:
          title: Dock Id
          type: string
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DockPatch'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DockAdminOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Patch Dock
      tags:
      - admin
  /api/admin/gateways:
    post:
      operationId: adminCreateGateway
      parameters:
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GatewayCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayCreatedOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Gateway
      tags:
      - admin
  /api/admin/gateways/pending/{gateway_id}:
    delete:
      operationId: adminDismissPending
      parameters:
      - in: path
        name: gateway_id
        required: true
        schema:
          title: Gateway Id
          type: string
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Dismiss Pending
      tags:
      - admin
  /api/admin/gateways/{gateway_id}:
    delete:
      operationId: adminDeleteGateway
      parameters:
      - in: path
        name: gateway_id
        required: true
        schema:
          title: Gateway Id
          type: string
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Gateway
      tags:
      - admin
    patch:
      operationId: adminPatchGateway
      parameters:
      - in: path
        name: gateway_id
        required: true
        schema:
          title: Gateway Id
          type: string
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GatewayPatch'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayPatchOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Patch Gateway
      tags:
      - admin
...
  /api/docks:
    get:
      operationId: listDocks
      parameters:
      - description: Filter by harbor
        in: query
        name: harbor_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by harbor
          title: Harbor Id
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/DockOut'
                title: Response Listdocks
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List all docks
      tags:
      - docks
  /api/docks/{dock_id}:
    get:
      operationId: getDock
      parameters:
      - in: path
        name: dock_id
        required: true
        schema:
          title: Dock Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DockWithBerthsOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a single dock with its berths
      tags:
      - docks
  /api/gateways:
    get:
      operationId: listGateways
      parameters:
      - description: Filter by harbor
        in: query
        name: harbor_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by harbor
          title: Harbor Id
      - description: Filter by dock
        in: query
        name: dock_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by dock
          title: Dock Id
      - description: Filter by gateway status
        in: query
        name: status
        required: false
        schema:
          anyOf:
          - pattern: ^(online|offline)$
            type: string
          - type: 'null'
          description: Filter by gateway status
          title: Status
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/GatewayOut'
                title: Response Listgateways
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - APIKeyCookie: []
      summary: List gateways visible to the harbormaster
      tags:
      - gateways
...
  /api/gateways/pending/{gateway_id}:
    delete:
      operationId: dismissPendingGateway
      parameters:
      - in: path
        name: gateway_id
        required: true
        schema:
          title: Gateway Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - APIKeyCookie: []
      summary: Dismiss a pending gateway entry
      tags:
      - gateways
...
servers:
- description: Local development
  url: http://localhost:8000
- description: Staging
  url: https://staging.example.com
...